changed the internal rep of movable from float to int

+++ include/Movable.h
@@ -1,35 +1,34 @@
 #pragma once
-#include <glm/vec2.hpp>
-
+#include <glm/glm.hpp>
 
 class Movable
 {
   public:
-    explicit Movable(glm::vec2 position) : m_position{position},m_velocity{}
+    explicit Movable(glm::ivec2 position) : m_position{position},m_velocity{}
     {
     }
 
-    [[nodiscard]]glm::vec2 position() const
+    [[nodiscard]]glm::ivec2 position() const
     {
         return m_position;
     }
 
-    [[nodiscard]] glm::vec2 velocity()const
+    [[nodiscard]] glm::ivec2 velocity()const
     {
         return m_velocity;
     }
 
-    void setVelocity(glm::vec2 newVelocity)
+    void setVelocity(glm::ivec2 newVelocity)
     {
         m_velocity = newVelocity;
     }
 
-    void setPosition(glm::vec2 newPosition)
+    void setPosition(glm::ivec2 newPosition)
     {
         m_position = newPosition;
     }
 
   private:
-    glm::vec2 m_position;
-    glm::vec2 m_velocity;
+    glm::ivec2 m_position;
+    glm::ivec2 m_velocity;
 };
